Rework bitmask in checkResourcesForNUMANodes
This commit rework the way we're working with bitmask in
checkResourcesForNUMANodes, previous approach (not the first one)
was in AND operation for bitmask which was set to ^uint64(0), that
approach didn't require accuracy in bitmask initialization, but new
one requires.

Also it adds unit test for case when we can't work with node resource
topology.

Signed-off-by: Alexey Perevalov <alexey.perevalov@huawei.com>