Add unit/system test for Scoring plugin
This test run a scenario where multiple pod requests arrived sequentially, and check if the scoring plugin ranks the available nodes as expected per each request.
The cretiria for the expected node, is based on the scoring model which is "Least Allocatable".

In the future I will add support for testing other scoring models as well.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>