nrt: logr: logID -> pod
Everywhere else in the kubernetes ecosystem, and most
notably in the scheduler, we call the pod namespace/name pair "pod",
while we call it "logID". We do it to use the same name for all
the flows, being the cache resync (which is driven by time, not
by an object) the odd one.

It seems better to be externally consistent (with the ecosystem)
rather than internally consistent (all the flows in the same plugin),
so we rename "logID" to "pod" in the log entries.

Along the way, replaced the inline constants with well known
key name constants from the logging (sub)package.

Signed-off-by: Francesco Romani <fromani@redhat.com>